{{ cookiecutter.repo_name }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_name }}"
version = "0.1.0"
description = "{{ cookiecutter.description }}"
authors = ["{{ cookiecutter.full_name }} <{{ cookiecutter.email }}>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"
toml = "^0.10.2"
numpy = "^1.20.3"
pandas = "^1.2.4"
jupyterlab = "^3.0.16"
mkdocs = "^1.1.2"
mkdocstrings = "^0.15.1"
mkdocs-material = "^7.1.6"

[tool.poetry.scripts]
main = "{{ cookiecutter.project_slug }}.__main__:main"

[tool.poetry.dev-dependencies]
ptvsd = "^4.3.2"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
